Split BMO and Ironic upgrade tests
